pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "imgconvert"
version = "0.2.0"
dependencies = ["click", "Pillow", "pillow-heif"]

[project.scripts]
imgconvert = "script:main"

[tool.setuptools]
py-modules = ["script", "analyze", "utils"]